added replying to replies

+++ app/db.py
@@ -1,6 +1,5 @@
-# app/db.py
 from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, DateTime
-from sqlalchemy.orm import sessionmaker, relationship
+from sqlalchemy.orm import sessionmaker, relationship, backref
 from sqlalchemy.ext.declarative import declarative_base
 from datetime import datetime
 from passlib.context import CryptContext
@@ -11,7 +10,6 @@ SQLALCHEMY_DATABASE_URI = 'sqlite:///database.db'
 Base = declarative_base()
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")  # Using bcrypt for hashing
 
-# Define SQLAlchemy models
 class User(Base):
     __tablename__ = 'users'
     id = Column(Integer, primary_key=True)
@@ -26,19 +24,31 @@ class BlogPost(Base):
     content = Column(String(1000), nullable=False)
     author = Column(String(50), nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
+    
+    # Relationship with comments
+    comments = relationship('Comment', back_populates='post')
 
 class Comment(Base):
     __tablename__ = 'comments'
     id = Column(Integer, primary_key=True)
-    content = Column(String(500), nullable=False)
+    content = Column(String(255), nullable=False)
     author = Column(String(50), nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow)
     post_id = Column(Integer, ForeignKey('blog_posts.id'))
-    post = relationship("BlogPost", back_populates="comments")
+    parent_comment_id = Column(Integer, ForeignKey('comments.id'))
 
-BlogPost.comments = relationship("Comment", back_populates="post")
+    # Define relationship for replies
+    replies = relationship('Comment', 
+                           cascade='all, delete-orphan', 
+                           backref=backref('parent', remote_side=[id], uselist=False), 
+                           remote_side=[parent_comment_id],
+                           lazy=True)
 
-# Database class for interactions
+    # Define relationship with BlogPost
+    post = relationship('BlogPost', back_populates='comments')
+
+
+# Define Database class for interactions
 class Database:
     def __init__(self):
         engine = create_engine(SQLALCHEMY_DATABASE_URI)
@@ -47,7 +57,7 @@ class Database:
         self.session = Session()
 
     def add_user(self, email, username, password):
-        hashed_password = pwd_context.hash(password)  # Hashing the password
+        hashed_password = pwd_context.hash(password)
         user = User(email=email, username=username, password=hashed_password)
         self.session.add(user)
         self.session.commit()
@@ -64,7 +74,7 @@ class Database:
         return self.session.query(BlogPost).all()
     
     def get_post(self, post_id):
-        return self.session.query(BlogPost).filter(BlogPost.id == post_id).one_or_none()
+        return self.session.query(BlogPost).filter_by(id=post_id).one_or_none()
 
     def update_blog_post(self, post_id, title, content):
         post = self.get_post(post_id)
@@ -75,14 +85,36 @@ class Database:
             return post
         return None
 
-    def delete_comment(self, comment):
-        self.session.delete(comment)
-        self.session.commit()
+    def delete_blog_post(self, post_id):
+        post = self.get_post(post_id)
+        if post:
+            self.session.delete(post)
+            self.session.commit()
+            return True
+        return False
 
-    def add_comment(self, post_id, content, author):
-        comment = Comment(content=content, author=author, post_id=post_id)
+    def add_comment(self, post_id, content, author, parent_comment_id=None):
+        comment = Comment(content=content, author=author, post_id=post_id, parent_comment_id=parent_comment_id)
         self.session.add(comment)
         self.session.commit()
 
+    def get_comments_for_post(self, post_id):
+        return self.session.query(Comment).filter_by(post_id=post_id, parent_comment_id=None).all()
+
+    def get_replies_for_comment(self, comment_id):
+        return self.session.query(Comment).filter_by(parent_comment_id=comment_id).all()
+
     def get_comment(self, post_id, comment_id):
         return self.session.query(Comment).filter_by(id=comment_id, post_id=post_id).one_or_none()
+
+    def delete_comment(self, comment):
+    # Ensure comment is an instance of Comment or fetch it by ID if necessary
+        if not isinstance(comment, Comment):
+            comment = self.session.query(Comment).filter_by(id=comment).first()
+
+        if comment:
+            self.session.delete(comment)
+            self.session.commit()
+            return True
+        return False
+
